add task solution
