--- src/components/ChangeTodoForm/index.ts
@@ -1,1 +1,0 @@
-export * from './ChangeTodoForm';
